+++ src/server/routes/iframeArticleRoute.js
@@ -15,6 +15,7 @@ import { getHtmlLang, getLocaleObject } from '../../i18n';
 import { fetchArticle } from '../../containers/ArticlePage/articleApi';
 import { fetchResourceTypesForResource } from '../../containers/Resources/resourceApi';
 import IframePage from '../../iframe/IframePage';
+import IframePageWrapper from '../../iframe/IframePageWrapper';
 import config from '../../config';
 import handleError from '../../util/handleError';
 import { renderPage, renderHtml } from '../helpers/render';
@@ -41,7 +42,13 @@ const getAssets = () => ({
 });
 
 function doRenderPage(initialProps) {
-  const Page = config.disableSSR ? '' : <IframePage {...initialProps} />;
+  const Page = config.disableSSR ? (
+    ''
+  ) : (
+    <IframePageWrapper {...initialProps}>
+      <IframePage {...initialProps} />
+    </IframePageWrapper>
+  );
   const { html, ...docProps } = renderPage(Page, getAssets(), {
     initialProps,
   });
